+++ kaparoo/filesystem/search/__init__.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 __all__ = (
     "SPEC_FILE_SUFFIXES",
+    "SearchKwargs",
     "Selector",
     "is_spec_file",
     "resolve_selector",
@@ -18,4 +19,5 @@ from kaparoo.filesystem.search.selection import (
     resolve_selector,
     select,
 )
+from kaparoo.filesystem.search.types import SearchKwargs
 from kaparoo.filesystem.search.wrappers import search_dirs, search_files, search_paths